fix: harden activity classification per code review
- tail_contents reads the live grid (scrollback offset zeroed during the
  read) so a user scrolled into history can't pin stale Working/Waiting
  markers or hide a fresh prompt (B1)
- trackers keyed by a never-reused Session::id instead of Arc pointer
  addresses, closing the kill/respawn key-collision window (B2)
- bell counter resyncs if it ever goes backwards (parser reset) instead
  of silently swallowing bells forever (B3)
- waiting patterns require menu structure (caret / numbered options) to
  co-occur with question phrases, so agent response prose mentioning
  'Do you want' no longer false-positives a dock bounce
- plain terminals never reach Done — typing echo isn't work

+++ src/gui/state.rs
@@ -121,10 +121,10 @@ pub struct Grove {
     /// input buffer. `view()` runs every tick; without this the modal would hit
     /// the filesystem (`read_dir`) on every frame.
     pub dir_cache: std::cell::RefCell<Option<(String, Vec<String>)>>,
-    /// Per-session activity trackers, keyed by the session's `dirty` Arc
-    /// pointer (same stable key as `pty_cache`). Refreshed every ~480ms by
-    /// `Msg::Tick`; stale keys are pruned on the same pass.
-    pub activity: HashMap<usize, super::activity::Tracker>,
+    /// Per-session activity trackers, keyed by `Session::id` (never reused,
+    /// unlike Arc pointer addresses). Refreshed every ~480ms by `Msg::Tick`;
+    /// stale keys are pruned on the same pass.
+    pub activity: HashMap<u64, super::activity::Tracker>,
     /// Whether the OS window currently has focus — gates the dock bounce.
     pub window_focused: bool,
     /// Last dock badge value pushed, to avoid redundant objc calls.
